Make Mario check if he's already been downloaded.

+++ ItsAMeMario.sh
@@ -1,6 +1,13 @@
 #!/usr/bin/env bash
 
-git clone --depth 8 https://github.com/JoshuaKGoldberg/Old-Deleted-FullScreenMario ~/.fullScreenMario
-cd ~/.fullScreenMario
-git reset --hard 77b493c9ce9fb015a88995658b11d95e28f1e68e
-xdg-open ./index.html
+FULL_SCREEN_MARIO_DIR=~/.fullScreenMario
+FULL_SCREEN_MARIO_FILE="$FULL_SCREEN_MARIO_DIR/index.html"
+
+if [[ -f "$FULL_SCREEN_MARIO_FILE" ]]; then
+  xdg-open "$FULL_SCREEN_MARIO_FILE"
+else
+  git clone --depth 8 https://github.com/JoshuaKGoldberg/Old-Deleted-FullScreenMario "$FULL_SCREEN_MARIO_DIR"
+  cd "$FULL_SCREEN_MARIO_DIR"
+  git reset --hard 77b493c9ce9fb015a88995658b11d95e28f1e68e
+  xdg-open "$FULL_SCREEN_MARIO_FILE"
+fi
